refactor: rename http to rest
