+++ app/transport/rest/middleware.go
@@ -1,4 +1,4 @@
-package http
+package rest
 
 import (
 	"net/http"
