+++ app/transport/rest/server.go
@@ -1,4 +1,4 @@
-package http
+package rest
 
 import (
 	"fmt"
